ecom_recommender_sv/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14,<4.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.2.0
mysqlclient>=2.0.0
gunicorn>=20.1.0
